Refactor resolve_item to separate the handling of traits from methods
